# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Poc para estudos de GO",
        "title": "Todo list",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/login": {
            "post": {
                "description": "Gera o token de acesso do usuário",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "description": "Informações do acesso",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UserLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Token"
                        }
                    },
                    "400": {
                        "description": "Parâmetro inválidos",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Acesso não authorizado",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Criar novo usuário",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "description": "Informações do usuário",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.User"
                        }
                    },
                    "400": {
                        "description": "Parâmetro inválidos",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/todo": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Listagem de todos cadatrados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todo"
                ],
                "summary": "Lista os todos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Todo"
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Criar novo todo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todo"
                ],
                "parameters": [
                    {
                        "description": "Payload que será criado",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Todo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Todo"
                        }
                    },
                    "400": {
                        "description": "Request inválido",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    },
                    "401": {
                        "description": "Token inválido",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    }
                }
            }
        },
        "/todo/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Atualizar um todo existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todo"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Identificador do todo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload que será atualizado",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Todo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Atualizado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    },
                    "400": {
                        "description": "Payload inválido",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Apaga um todo pelo identificado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todo"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Identificado único do todo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Error message"
                },
                "message": {
                    "type": "string",
                    "example": "Message"
                }
            }
        },
        "domain.CreateUser": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "jhon.doel@example.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "Jhon"
                },
                "last_name": {
                    "type": "string",
                    "example": "Doel"
                },
                "password": {
                    "type": "string",
                    "example": "s3cr3t3"
                },
                "username": {
                    "type": "string",
                    "example": "jhondoel1995"
                }
            }
        },
        "domain.Todo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Example description"
                },
                "done": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string",
                    "example": "1"
                },
                "name": {
                    "type": "string",
                    "example": "Example name"
                }
            }
        },
        "domain.Token": {
            "type": "object",
            "properties": {
                "expires_in": {
                    "type": "integer"
                },
                "token": {
                    "type": "string",
                    "example": "asdfas-asdfasd-asdf-asdf-asdf"
                }
            }
        },
        "domain.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@email.com"
                },
                "enabled": {
                    "type": "boolean"
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string",
                    "example": "Doel"
                },
                "username": {
                    "type": "string",
                    "example": "jhon_doel"
                }
            }
        },
        "domain.UserLogin": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "s3cr3t3"
                },
                "username": {
                    "type": "string",
                    "example": "jhondoel"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}